requirements.txt
django
djangae
oauth2client==1.5.2
git+https://github.com/mozilla/django-csp.git#egg=djangocsp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
six
djangorestframework
django-extensions
bpython
